requirements.txt
asttokens==2.4.1
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.29.5
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msrest==0.7.1
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1